rtnetlink: Add public interface parser
This will be used by the rtnetlink server to parse new links.
