+++ src/rtnetlink.hpp
@@ -8,6 +8,8 @@
 namespace phosphor::network::netlink
 {
 
+InterfaceInfo intfFromRtm(std::string_view msg);
+
 std::optional<std::tuple<unsigned, InAddrAny>>
     gatewayFromRtm(std::string_view msg);
