+++ src/system_queries.hpp
@@ -42,7 +42,6 @@ struct NeighborFilter
 
 namespace detail
 {
-InterfaceInfo parseInterface(const nlmsghdr& hdr, std::string_view msg);
 bool validateNewAddr(const AddressInfo& info,
                      const AddressFilter& filter) noexcept;
 bool validateNewNeigh(const NeighborInfo& info,
